- Use long longs instead of ints when looking at table sizes, free space, and row counts for display, fixing overflow issues (this addresses Issue #394)  - Tweak the byte size formatter to also show TBs for very large numbers  - Improve row count and auto increment value formatting by using a number formatter (enhancements for Issue #394)

+++ Source/SPStringAdditions.h
@@ -40,7 +40,7 @@ static inline const char* NSStringUTF8String(NSString* self) {
 
 @interface NSString (SPStringAdditions)
 
-+ (NSString *)stringForByteSize:(int)byteSize;
++ (NSString *)stringForByteSize:(long long)byteSize;
 + (NSString *)stringForTimeInterval:(float)timeInterval;
 
 - (NSString *)backtickQuotedString;
